camera: the coverity of android U camera hal [1/1]
PD#SWPL-123668

Problem:
The code of the android U camera hal have
coverity problem

Solution:
Fix coverity problem

Verify:
android U adt4

+++ camera/v3/fake-pipeline2/OMXDecoder.h
@@ -138,7 +138,7 @@ public:
     }
 private:
     int decoderType;
-    int mWaitVsyncDuration;
+    int64_t mWaitVsyncDuration;
 
     OMX_ERRORTYPE WaitForState(OMX_HANDLETYPE hComponent, OMX_STATETYPE eTestState, OMX_STATETYPE eTestState2);
     OMX_U32 mInWidth;
